requirements.txt
torch==1.10
numpy==1.24.4
pyyaml